requirements.txt
pymobiledevice3==9.12.0
reportlab>=4.0.0,<4.4.0
jinja2>=3.0.0
customtkinter>=5.0.0
Pillow>=9.0.0
